This is Version 3
A couple things.
I took out the background image entriely and picked a color I liked for the whole background of the site.
I changed the mobile site to have a dropdown menu as well as changed slider to just pictures so they can still go to github and find my work
Added an about me to my about page.
Took out the email button to a link on the landing page,
Changed the heading of the website
Changed the navigation slightly and tried my best to fix navigation issues on mobile.
Logo for Version 3 coming soon.